+++ lib/ui/editor_view.dart
@@ -81,24 +81,29 @@ class _EditorViewState extends State<EditorView> {
 
   Future<void> setColor() async {
     if (!isColorSet) {
-      final bool containsImage = widget.controller.assets.value.any((StoryElement element) {
+      final bool containsImage =
+          widget.controller.assets.value.any((StoryElement element) {
         return element.type == ItemType.image || element.type == ItemType.video;
       });
       if (containsImage) {
         final StoryElement element =
             widget.controller.assets.value.firstWhere((StoryElement element) {
-          return element.type == ItemType.image || element.type == ItemType.video;
+          return element.type == ItemType.image ||
+              element.type == ItemType.video;
         });
         if (element.type == ItemType.video) {
-          final CoverData videoCover = await generateSingleCoverThumbnail(element.value);
-          paletteColor =
-              await PaletteGeneratorUtil.getGeneratorFromData(videoCover.thumbData ?? Uint8List(0));
+          final CoverData videoCover =
+              await generateSingleCoverThumbnail(element.value);
+          paletteColor = await PaletteGeneratorUtil.getGeneratorFromData(
+              videoCover.thumbData ?? Uint8List(0));
         } else {
-          paletteColor = await PaletteGeneratorUtil.getGeneratorFromPath(element.value);
+          paletteColor =
+              await PaletteGeneratorUtil.getGeneratorFromPath(element.value);
         }
         isColorSet = true;
         if (paletteColor != null) {
-          widget.controller.storyModel.paletteColors = paletteColor!.colors.toList();
+          widget.controller.storyModel.paletteColors =
+              paletteColor!.colors.toList();
           setState(() {});
         }
       }
@@ -125,7 +130,8 @@ class _EditorViewState extends State<EditorView> {
                   decoration: BoxDecoration(
                     gradient: LinearGradient(
                       colors: <Color>[
-                        if (paletteColor != null && paletteColor!.colors.isNotEmpty) ...<Color>[
+                        if (paletteColor != null &&
+                            paletteColor!.colors.isNotEmpty) ...<Color>[
                           paletteColor!.colors.first,
                           paletteColor!.colors.last
                         ] else ...<Color>[
@@ -146,7 +152,8 @@ class _EditorViewState extends State<EditorView> {
                       ),
                       ValueListenableBuilder<ColorFilterGenerator>(
                         valueListenable: widget.controller.selectedFilter,
-                        builder: (BuildContext context, ColorFilterGenerator value, Widget? child) {
+                        builder: (BuildContext context,
+                            ColorFilterGenerator value, Widget? child) {
                           return ColorFiltered(
                             colorFilter: ColorFilter.matrix(value.matrix),
                             child: child,
@@ -174,7 +181,8 @@ class _EditorViewState extends State<EditorView> {
                       ),
                       ValueListenableBuilder<bool>(
                         valueListenable: widget.controller.isShowingOverlay,
-                        builder: (BuildContext context, bool value, Widget? child) {
+                        builder:
+                            (BuildContext context, bool value, Widget? child) {
                           return Positioned(
                             top: 0,
                             width: MediaQuery.of(context).size.width,
@@ -194,7 +202,8 @@ class _EditorViewState extends State<EditorView> {
                       ),
                       ValueListenableBuilder<bool>(
                         valueListenable: widget.controller.isShowingOverlay,
-                        builder: (BuildContext context, bool value, Widget? child) {
+                        builder:
+                            (BuildContext context, bool value, Widget? child) {
                           return Positioned(
                             right: 0,
                             height: MediaQuery.of(context).size.height,
@@ -205,7 +214,8 @@ class _EditorViewState extends State<EditorView> {
                                   : widget.actionsBar ??
                                       ActionsBarWidget(
                                         editorController: widget.controller,
-                                        additionalActions: widget.additionalActions,
+                                        additionalActions:
+                                            widget.additionalActions,
                                       ),
                             ),
                           );